feat: Add support for formatting and retrieving Hijri dates in Arabic
This commit adds a new static method `Date` to the `Hijirdatetime` class in order to support formatting and retrieving Hijri dates in Arabic. The method takes a format string and an optional timestamp and returns the Hijri date formatted according to the given format in Arabic.

The code changes include adding the `Date` method to the `Hijirdatetime` class and updating the `examples.php` file to demonstrate the usage of the new method.